Development

+++ app/Repositories/PlayerRepository.php
@@ -7,41 +7,49 @@ use App\Models\Player;
 
 class PlayerRepository implements PlayerInterface
 {
-    protected $player;
-    
-    public function __construct(Player $player)
-    {
-        $this->player = $player;
-    }
-    
-    public function index()
-    {
-        return $this->player::with('class')->get();
-    }
-    
-    public function store($data)
-    {
-        return $this->player::create($data);
-    }
-    
-    public function show($id)
-    {
-        return $this->player::with('class')->findOrFail($id);
-    }
-    
-    public function update($data, $id)
-    {
-        $player = $this->player::findOrFail($id);
-        $player->update($data);
+    protected Player $playerModel;
 
+    public function __construct(Player $playerModel)
+    {
+        $this->playerModel = $playerModel;
+    }
+
+    public function getAll()
+    {
+        return $this->playerModel->with('class')->get();
+    }
+
+    public function getById($id)
+    {
+        return $this->playerModel->with('class')->find($id);
+    }
+
+    public function create($data)
+    {
+        return $this->playerModel->create($data);
+    }
+
+    public function update($id, array $data)
+    {
+        $player = $this->getById($id);
+
+        if (!$player) {
+            return null;
+        }
+
+        $player->update($data);
         return $player;
     }
-    
-    public function destroy($id)
-    {
-        $player = $this->player::findOrFail($id);
-        $player->delete();
 
+    public function delete($id)
+    {
+        $player = $this->getById($id);
+
+        if (!$player) {
+            return null;
+        }
+
+        $player->delete();
         return $player;
     }
 }
